+++ src/types.ts
@@ -12,6 +12,10 @@ export interface RestUrl {
    * GET请求时合并到自动合并到请求参数中
    */
   query?: Record<string, string | number>;
+  /**
+   * 实例中的restApi地址会覆盖全局地址,key必须大写
+   */
+  restApi?:Record<string, string|{url:string,method:string}>;
   [param: string]: unknown;
 }